More sophisticated dynamic encoding support
Added various functions for manually encoding APs (but not all of the functions
that would eventually be needed). If we do go this route, then we should create
a generic "add_repeated_varint() function, which takes a pointer to a function
which knows how to obtain the specific varints to be encoded (e.g., macs, rssi,
etc.).