[ELG-241] Log state buffer in sky_open and sky_close
Add logging of CRC of state buffer when it is saved during sky_close and
when it is restored in sky_open. The CRCs should match if the state is
successfully saved and then restored.
Avoid storing pointer in state by changing field 'newest' from a cacheline pointer to
a simple index into the cacheline array.
Add conditional compilation around logging in utility.c
Use printf in unit_test rather than LOGFMT so these logs are printed
when SKY_DEBUG is false