[EC-160] fix off by one for loop.
This bug was very unlikely to affect the score. The bug causes an access to beacon array at the index
one before the first cell. This may be at index -1 or to the last AP. Either way, the memory is compared
to beacons in the cache. If it is an AP, we have already established that it does not match, if it is
the array index -1, then it is extremely unlikely that it is a match to any cell in the cache. This means
that the score which is based on cells that do match is unaffected.