Example in sample client needs updating due to apparent change in API server db
The sample client has an example scan that should not match cache due to
the presence of gnss, however, the scan does not cause a cache hit
without the gnss! I have added an additional AP and cell to the scan so that
there is now a potential cache hit.